refactor to process system message like any other message
